requirements.txt
fastapi==0.111.0
pydantic-settings==2.2.1
pytz==2019.3
requests==2.32.3
